Trigram model with discounting and backing off
This code will build unigrams, bigrams, and trigrams language models. For bigrams and trigrams, I applied absolute discount and backing off to handle zero probabilities. The code hasn't been optimized yet, so it still takes quite a long time to build the models and consumes a lot of memory. At the end of the program, it will output metadata file, which contains vocabulary size and words' ids, and three language models.